Readded Sounds to Projectiles
Readded sounds to explosive projectiles, also continued cleaning up the code base. Began fixing the sounds for the Nuke, Phase bomb, and C4 although none of them are working 100% atm
Sounds are back in the game but some of the aren't quite right, likely due to pitch and volume tweaks being needed.

+++ Projectiles/PhaseBombProjectile.cs
@@ -17,12 +17,9 @@ namespace ExtraExplosives.Projectiles
         //private Mod ThoriumMod = ModLoader.GetMod("ThoriumMod");
         internal static bool CanBreakWalls;
         private SoundStyle phaseSound;
-        private ReLogic.Utilities.SlotId phaseSoundInstance;
 
         public override void SetStaticDefaults()
         {
-            // DisplayName.SetDefault("PhaseBomb");
-            //Tooltip.SetDefault("Your one stop shop for all your turretaria needs.");
             Main.projFrames[Projectile.type] = 10;
         }
 
@@ -58,19 +55,20 @@ namespace ExtraExplosives.Projectiles
         {
             //projectile.tileCollide = false;
 
-            //if (phaseSoundInstance == null)
-                //phaseSoundInstance = //SoundEngine.PlaySound(phaseSound);
+            // Not needed ?
+            /*if (SoundEngine.TryGetActiveSound(phaseSound.))
+                phaseSound = SoundEngine.PlaySound(phaseSound);*/
 
-            /* TODO if (phaseSoundInstance.State != SoundState.Playing)
-                phaseSoundInstance.Play();*/
+            if (!SoundEngine.FindActiveSound(phaseSound).IsPlaying)
+                SoundEngine.PlaySound(phaseSound);
 
             //Player player = Main.player[Projectile.owner];
 
-            /* TODO if (!Main.mouseLeft && Projectile.timeLeft <= 940)
+            if (!Main.mouseLeft && Projectile.timeLeft <= 940)
             {
-                phaseSoundInstance.Stop(true);
+                SoundEngine.FindActiveSound(phaseSound).Stop();
                 Projectile.Kill();
-            }*/
+            }
 
 
             if (++Projectile.frameCounter >= 5)
